% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolCorrApprox.R
\name{bolCorrApprox}
\alias{bolCorrApprox}
\title{Calculate the bolometric correction approximately for a star, given its temperature.}
\usage{
bolCorrApprox(temperature)
}
\arguments{
\item{temperature}{The star's effective temperature in Kelvin}
}
\value{
Bolometric Correction (BC)
}
\description{
Calculate the bolometric correction approximately for a star, given its temperature.
}
\examples{
bolCorrApprox(2800)

}
