% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starDist.R
\name{starDist}
\alias{starDist}
\title{Calculate distance to a (main sequence) star}
\usage{
starDist(relative_luminosity, absolute_magnitude)
}
\arguments{
\item{relative_luminosity}{Relative luminosity of the star (L/Lo)}

\item{absolute_magnitude}{Absolute magnitude of the star (Mv)}
}
\value{
Distance to the star (in parsecs)
}
\description{
Calculate distance to a (main sequence) star
}
\examples{
starDist(0.0024, 16.1)

}
