% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starDistMainSeq.R
\name{starDistMainSeq}
\alias{starDistMainSeq}
\title{Calculate distance to a (main sequence) star}
\usage{
starDistMainSeq(relative_luminosity, absolute_magnitude)
}
\arguments{
\item{relative_luminosity}{Relative luminosity of the star (L/Lo)}

\item{absolute_magnitude}{Absolute magnitude of the star (Mv)}
}
\value{
Distance to the main sequence star (in parsecs)
}
\description{
Calculate distance to a (main sequence) star
}
\examples{
starDistMainSeq(0.0024, 16.1)

}
