% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSurfaceArea.R
\name{starSurfaceArea}
\alias{starSurfaceArea}
\title{Calculate the surface area of the star in kilometers}
\usage{
starSurfaceArea(relative_radius)
}
\arguments{
\item{relative_radius}{Relative Radius of the star (R/Ro) Where Ro is the radius of the Sun}
}
\value{
Surface Area of the star (in kilometers)
}
\description{
Calculate the surface area of the star in kilometers
}
\examples{
starSurfaceArea(0.16)

}
